Move timeDifference util to shared
