+++ athena/utils/push-notifications/notification-formatting.js
@@ -1,6 +1,6 @@
 import onlyContainsEmoji from 'shared/only-contains-emoji';
 import sentencify from 'shared/sentencify';
-import { short as timeDifferenceShort } from 'shared/time-difference';
+import { timeDifferenceShort } from 'shared/time-difference';
 import sortByDate from 'shared/sort-by-date';
 import { toState, toPlainText } from 'shared/draft-utils';
