+++ shared/time-difference.js
@@ -1,26 +1,17 @@
-/**
- * This file is shared between server and client.
- * ⚠️ DON'T PUT ANY NODE.JS OR BROWSER-SPECIFIC CODE IN HERE ⚠️
- *
- * Note: This uses Flow comment syntax so this whole file is actually valid JS without any transpilation
- * The reason I did that is because create-react-app doesn't transpile files outside the source folder,
- * so it chokes on the Flow syntax.
- * More info: https://flow.org/en/docs/types/comments/
- */
+// @flow
+const MS_PER_SECOND = 1000;
+const MS_PER_MINUTE = 60000;
+const MS_PER_HOUR = 3600000;
+const MS_PER_DAY = 86400000;
+const MS_PER_YEAR = 31536000000;
 
-var MS_PER_SECOND = 1000;
-var MS_PER_MINUTE = 60000;
-var MS_PER_HOUR = 3600000;
-var MS_PER_DAY = 86400000;
-var MS_PER_YEAR = 31536000000;
-
-function timeDifferenceShort(current /*: Date*/, previous /*: Date*/) {
-  var elapsed = current - previous;
+export function timeDifferenceShort(current: Date, previous: Date) {
+  const elapsed = current - previous;
 
   if (elapsed < MS_PER_MINUTE) {
     return Math.round(elapsed / MS_PER_SECOND) + 's';
   } else if (elapsed < MS_PER_HOUR) {
-    return Math.round(elapsed / msPerMinute) + 'm';
+    return Math.round(elapsed / MS_PER_MINUTE) + 'm';
   } else if (elapsed < MS_PER_DAY) {
     return Math.round(elapsed / MS_PER_HOUR) + 'h';
   } else if (elapsed < MS_PER_YEAR) {
@@ -30,6 +21,59 @@ function timeDifferenceShort(current /*: Date*/, previous /*: Date*/) {
   }
 }
 
-module.exports = {
-  short: timeDifferenceShort,
-};
+export function timeDifference(current: number, previous: ?number): string {
+  if (!previous) return '';
+
+  const msPerMinute = 60 * 1000;
+  const msPerHour = msPerMinute * 60;
+  const msPerDay = msPerHour * 24;
+  const msPerMonth = msPerDay * 30;
+  const msPerYear = msPerDay * 365;
+
+  let elapsed = current - previous;
+
+  if (elapsed < msPerMinute) {
+    return 'Just now';
+  } else if (elapsed < msPerHour) {
+    const now = Math.round(elapsed / msPerMinute);
+    if (now === 1) {
+      return '1 minute ago';
+    } else {
+      return `${now} minutes ago`;
+    }
+  } else if (elapsed < msPerDay) {
+    const now = Math.round(elapsed / msPerHour);
+    if (now === 1) {
+      return '1 hour ago';
+    } else {
+      return `${now} hours ago`;
+    }
+  } else if (elapsed < msPerMonth) {
+    const now = Math.round(elapsed / msPerDay);
+    if (now === 1) {
+      return 'Yesterday';
+    } else if (now >= 7 && now <= 13) {
+      return '1 week ago';
+    } else if (now >= 14 && now <= 20) {
+      return '2 weeks ago';
+    } else if (now >= 21 && now <= 28) {
+      return '3 weeks ago';
+    } else {
+      return `${now} days ago`;
+    }
+  } else if (elapsed < msPerYear) {
+    const now = Math.round(elapsed / msPerMonth);
+    if (now === 1) {
+      return '1 month ago';
+    } else {
+      return `${now} months ago`;
+    }
+  } else {
+    const now = Math.round(elapsed / msPerYear);
+    if (now === 1) {
+      return '1 year ago';
+    } else {
+      return `${now} years ago`;
+    }
+  }
+}
