+++ src/views/thread/components/threadDetail.js
@@ -6,9 +6,9 @@ import { withRouter } from 'react-router';
 import Link from 'src/components/link';
 import {
   getLinkPreviewFromUrl,
-  timeDifference,
   convertTimestampToDate,
 } from '../../../helpers/utils';
+import { timeDifference } from 'shared/time-difference';
 import isURL from 'validator/lib/isURL';
 import { URLS } from '../../../helpers/regexps';
 import { openModal } from '../../../actions/modals';
